requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
boto3>=1.26.137
pydantic>=2.7.0
pydantic-settings>=2.7.1
streamlit>=1.32.0
Pillow>=9.5.0
gTTS>=2.3.1
requests>=2.31.0
python-multipart>=0.0.6
pygame>=2.1.0